+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+# 0.2.2 - Unreleased
+
+### Fixed
+
+- Support inline eval of macros with `#` and `##` syntax
+
 # 0.2.1 - 2025-03-23
 
 ### Added